Support for limit plyaback range on Actions
Added support for the "Limit Playback Range" option in the exporter. Using this will crop actions in the NLA editor to the size of the used strip instead of the full range of keyframes.
It will also use the current set start and end frame as crop borders for the active Action.